agent.c becomes two files: one TU per platform
Deferred three times as ceremony, the split is mechanical
now -- cycles v2.64-v2.66 took the policy ladder, the frame
codec, and the ring out, and what remained shared a TU and a
diff budget while sharing nothing else: own state structs,
own I/O, own threading, one #else at line 1073. Each half
becomes its own file; CMake picks per-platform exactly as
tools/retraced does for its transports. No behavior change;
the shared API (agent.h) and the shared modules are
untouched. Every future platform edit starts in the right
file, and each half compiles visibly on its own host -- the
preprocessor dead zones that hid type errors from the
compiling machine are gone.
